+++ example/src/forms/InlineValidatorsForm/InlineValidatorsForm.js
@@ -1,9 +1,10 @@
-import Form, { Input, validators } from '@detools/vue-form'
+import Form, { Input, Notification, validators } from '@detools/vue-form'
 
 export default {
   data() {
     return {
       formValues: {},
+      formErrors: {},
     }
   },
 
@@ -12,15 +13,41 @@ export default {
       this.formValues = values
     },
 
-    asyncValidator(value, name) {
+    handleReset() {
+      this.formValues = {}
+    },
+
+    handleDisabled(errors) {
+      Notification.warning('handleDisabled method has been called')
+
+      this.formErrors = errors
+    },
+
+    asyncValidator3000(value, name) {
       return new Promise((resolve, reject) => {
         setTimeout(() => {
           if (value === 'github') {
+            Notification.success(`${name} passed a validation (3s)`)
             resolve()
           } else {
-            reject(`${name} should be "github"`)
+            Notification.error(`${name} did not pass a validation (3s)`)
+            reject(`${name} should be github`)
           }
-        }, 2000)
+        }, 3000)
+      })
+    },
+
+    asyncValidator5000(value, name) {
+      return new Promise((resolve, reject) => {
+        setTimeout(() => {
+          if (value) {
+            Notification.success(`${name} passed a validation (5s)`)
+            resolve()
+          } else {
+            Notification.error(`${name} did not pass a validation (5s)`)
+            reject(`${name} is required`)
+          }
+        }, 5000)
       })
     },
   },
@@ -31,13 +58,25 @@ export default {
         <h1>Inline Validators Form</h1>
         <div class="wrapper">
           <div class="form">
-            <Form reset save submit handleSubmit={this.handleSubmit}>
+            <Form
+              reset
+              save
+              submit
+              handleSubmit={this.handleSubmit}
+              handleReset={this.handleReset}
+              handleDisabled={this.handleDisabled}>
               <Input
                 formItem
-                name="asyncUsername"
-                label="Username"
+                name="username"
+                label="Username with async validator for 3s"
                 validators={[validators.isRequired(), validators.length({ min: 6 })]}
-                asyncValidators={[this.asyncValidator]}
+                asyncValidators={[this.asyncValidator3000]}
+              />
+              <Input
+                formItem
+                name="anything"
+                label="Anything with async validator for 5s"
+                asyncValidators={[this.asyncValidator5000]}
               />
             </Form>
           </div>
@@ -47,6 +86,14 @@ export default {
             <br />
             <div>
               <pre>{JSON.stringify(this.formValues, null, 2)}</pre>
+            </div>
+            <br />
+            <br />
+            <strong>Form Errors</strong>
+            <br />
+            <br />
+            <div>
+              <pre>{JSON.stringify(this.formErrors, null, 2)}</pre>
             </div>
           </div>
         </div>